+++ lib/paypal.js
@@ -52,7 +52,7 @@ Meteor.Paypal = {
     return options;
   },
   authorize: function(cardInfo, paymentInfo, callback) {
-    Meteor.call("paypalSubmit", "authorize", cardInfo, paymentInfo, callback);
+    Meteor.call("payflowProSubmit", "authorize", cardInfo, paymentInfo, callback);
   },
   capture: function(transactionId, amount, callback) {
     var captureDetails;
